+++ packages/vda-sbt-client/test/blockchain-api.test.ts
@@ -1,16 +1,30 @@
 require('dotenv').config();
-import { dids, getBlockchainAPIConfiguration, getNetwork, CONTEXT_NAME, RECIPIENT_WALLET } from "./utils"
+import { dids, getBlockchainAPIConfiguration, getNetwork, getSignerInfo, CONTEXT_NAME, RECIPIENT_WALLET } from "./utils"
 import { VeridaSBTClient } from "../src/index"
 import { ethers, Wallet } from "ethers";
-import EncryptionUtils from "@verida/encryption-utils";
+import { Keyring } from "@verida/keyring";
 
 const assert = require('assert')
 
 const did = dids[0];
-  
 
-const configuration = getBlockchainAPIConfiguration();
-const createBlockchainAPI = (did: any) => {
+// tokenReceiver should be the same as the did to test the burn() function 
+// of the SBT contract. Because only the owner of tokenID can call the burn 
+// function and all the transactions of vda-sbt-client instance will be 
+// signed by the did
+const tokenReceiver = RECIPIENT_WALLET
+
+const mintedTokenIds = [1, 2, 3] //[1, 3, 5]
+const burntTokenIds =  [22] //[2, 4]
+
+interface InterfaceDID {
+    address: string
+    privateKey: string
+    publicKey: string
+}
+
+const createBlockchainAPI = (did: InterfaceDID) => {
+    const configuration = getBlockchainAPIConfiguration(did.privateKey);
     return new VeridaSBTClient({
         identifier: did.address,
         signKey: did.privateKey,
@@ -19,11 +33,90 @@ const createBlockchainAPI = (did: any) => {
     })
 }
 
+const claimSBT = async (
+    didAddress: string,
+    sbtType: string,
+    uniqueId: string,
+    sbtURI: string,
+    receiverAddress: string,
+    signedProof: string,
+    keyring: Keyring,
+    blockchainApi: VeridaSBTClient
+) => {
+    const signedDataMsg = ethers.utils.solidityPack(
+        ['string','address'],
+        [`${sbtType}-${uniqueId}-`, didAddress.toLowerCase()]
+    )
+    const signedData = await keyring.sign(signedDataMsg)
+
+    await blockchainApi.claimSBT(
+        sbtType,
+        uniqueId,
+        sbtURI,
+        receiverAddress.toLowerCase(),
+        signedData,
+        signedProof
+    )
+}
+
+const createTestDataIfNotExist = async (
+    trustedSigner: InterfaceDID, 
+    blockchainApi: VeridaSBTClient,
+    keyring: Keyring,
+    signedProof: string
+    ) => {
+    const orgTotalSupply = parseInt(await blockchainApi.totalSupply())
+    if (orgTotalSupply > 0) return;
+
+    console.log("Creating initial test data by mint & burn operations")
+
+    const sbtType = "twitter-test"
+    const uniqueIds = ["00001", "00002", "00003", "00004", "00005"]
+    const sbtURI = "https://gateway.pinata.cloud/ipfs/QmVrTkbrzNHRhmsh88XnwJo5gBu8WqQMFTkVB4KoVLxSEY/3.json"
+
+    
+    // Should check if the trustedSigner is registered to the contract
+    const signers = await blockchainApi.getTrustedSignerAddresses()
+    assert.ok(signers.includes(trustedSigner.address))
+
+    // Mint tokens
+    for (let i = 0; i < uniqueIds.length; i++) {
+        await claimSBT(
+            did.address, 
+            sbtType, 
+            uniqueIds[i], 
+            sbtURI, 
+            tokenReceiver.address.toLowerCase(),
+            signedProof, 
+            keyring, 
+            blockchainApi
+        )
+    }
+    const newTotalSupply = parseInt(await blockchainApi.totalSupply())
+
+    assert.ok(newTotalSupply === 5, "SBT claimed successfully")
+
+    // burn last 2 tokens
+    const receiverSBTClient = createBlockchainAPI(RECIPIENT_WALLET)
+    for (const id of burntTokenIds) {
+        await receiverSBTClient.burnSBT(id)
+    }
+
+    console.log("Initial test data created")
+}
+
 describe('vda-sbt-client blockchain api', () => {
     let blockchainApi : VeridaSBTClient
 
-    before(() => {
+    const trustedSigner = dids[2]
+    let keyring: Keyring
+    let signedProof: string
+
+    before(async () => {
         blockchainApi = createBlockchainAPI(did);
+        [keyring, signedProof] = await getSignerInfo(trustedSigner.privateKey)
+
+        await createTestDataIfNotExist(trustedSigner, blockchainApi, keyring, signedProof)
     })
 
     describe("totalSupply", () => {
@@ -67,16 +160,16 @@ describe('vda-sbt-client blockchain api', () => {
         })
 
         it("Should reject for burnt token IDs",async () => {
-            await assert.rejects(
-                blockchainApi.isLocked(22)
-            )
+            for (const tokenId of burntTokenIds) {
+                await assert.rejects(
+                    blockchainApi.isLocked(tokenId)
+                )
+            }
         })
 
         it("Should be locked for all minted tokenIDs",async () => {
-            const tokenIdList = [1,2,3,20]
-
-            for (let i = 0; i <= tokenIdList.length; i+=7) {
-                const locked = await blockchainApi.isLocked(tokenIdList[i])
+            for (const tokenId of mintedTokenIds) {
+                const locked = await blockchainApi.isLocked(tokenId)
                 assert.equal(locked, true, "Locked correctly")
             }
         })
@@ -93,7 +186,7 @@ describe('vda-sbt-client blockchain api', () => {
         this.timeout(60*1000)
 
         // SBT info
-        const sbtType = "twitter-test"
+        const sbtType = "twitters"
         const uniqueId = "12346789"
         const sbtURI = "https://gateway.pinata.cloud/ipfs/QmVrTkbrzNHRhmsh88XnwJo5gBu8WqQMFTkVB4KoVLxSEY/3.json"
 
@@ -107,36 +200,30 @@ describe('vda-sbt-client blockchain api', () => {
             const signers = await blockchainApi.getTrustedSignerAddresses()
             assert.ok(signers.includes(trustedSigner.address))
 
-            // Create SBT Info
-            const signedDataMsg = ethers.utils.solidityPack(
-                ['string','address'],
-                [`${sbtType}-${uniqueId}-`, did.address]
-            )
-            const signedData = await trustedSignerNetworkInfo.keyring.sign(signedDataMsg)
-            
-            // Fetch the context proof from the signing DID
-            const signedProof = await trustedSignerDIDDocument.locateContextProof(CONTEXT_NAME)
-
-            // Claim SBT
+            // Claim a SBT
             const orgTotalSupply = parseInt(await blockchainApi.totalSupply())
-
-            await blockchainApi.claimSBT(
+            await claimSBT(
+                did.address,
                 sbtType,
                 uniqueId,
                 sbtURI,
-                RECIPIENT_WALLET.address,
-                signedData,
-                signedProof!
+                RECIPIENT_WALLET.address.toLowerCase(),
+                signedProof,
+                keyring,
+                blockchainApi
             )
-
             const newTotalSupply = parseInt(await blockchainApi.totalSupply())
-
             assert.ok(newTotalSupply === (orgTotalSupply + 1), "SBT claimed successfully")
         })
 
         it("isSBTClaimed()",async () => {
-            const claimed = await blockchainApi.isSBTClaimed(sbtType, uniqueId)
+            const claimed = await blockchainApi.isSBTClaimed(RECIPIENT_WALLET.address, sbtType, uniqueId)
             assert.ok(claimed === true)
+        })
+
+        it("getClaimedSBTList()",async () => {
+            const idList = await blockchainApi.getClaimedSBTList(RECIPIENT_WALLET.address)
+            assert.ok(idList.length > 0)
         })
 
         it("tokenInfo()",async () => {
@@ -152,7 +239,8 @@ describe('vda-sbt-client blockchain api', () => {
             const owner = await blockchainApi.ownerOf(lastTokenId)
             assert.equal(owner, RECIPIENT_WALLET.address)
 
-            await blockchainApi.burnSBT(lastTokenId)
+            const receiverSBTClient = createBlockchainAPI(RECIPIENT_WALLET)
+            await receiverSBTClient.burnSBT(lastTokenId)
 
             assert.rejects(
                 blockchainApi.ownerOf(lastTokenId),